refactor(inventory): 발주 조회를 PurchaseOrderReader 로 분리 (#724 항목 5-b)
getPurchaseOrderById → findById, getPurchaseOrders → findMany. 본문 변경 0줄.
쓰기 경로가 응답을 만들려 이 리더를 6곳에서 부르므로 Service 가 리더를 주입받고,
컨트롤러가 쓰던 두 메서드는 위임으로 남긴다(포트 모양의 시작).

리더 docstring 에 두 가지를 박았다: 헤더 expectedArrival 이 컬럼이 아니라 ordered
라인 ETA 의 MIN 파생이라는 것, 그리고 이 파일에 .for('update') 를 넣는 순간 잠금
불변식(PO 행 → 라인 행)의 적용 대상이 되므로 Manager 로 옮기라는 것.

PurchaseOrderService: 678 → 517줄.

검증: type-check 0 · eslint 0 · jest 통과 · nest build 성공 · 통합 기준선과 동일
