+++ apps/core/src/modules/inventory/procurement/procurement.module.ts
@@ -6,6 +6,7 @@ import { PurchaseOrderController } from './controllers/purchase-order.controller
 import { PurchaseOrderService } from './services/purchase-order.service';
 import { PurchaseOrderCartService } from './services/purchase-order-cart.service';
 import { ReorderSuggestionReader } from './services/reorder-suggestion.reader';
+import { PurchaseOrderReader } from './services/purchase-order.reader';
 
 /**
  * 조달(발주) 모듈. 경계는 ADR-0032 가 소유한다.
@@ -25,7 +26,7 @@ import { ReorderSuggestionReader } from './services/reorder-suggestion.reader';
 @Module({
   imports: [CoreInventoryModule, SharedModule, InboundModule],
   controllers: [PurchaseOrderController],
-  providers: [PurchaseOrderService, PurchaseOrderCartService, ReorderSuggestionReader],
-  exports: [PurchaseOrderService, PurchaseOrderCartService, ReorderSuggestionReader],
+  providers: [PurchaseOrderService, PurchaseOrderCartService, ReorderSuggestionReader, PurchaseOrderReader],
+  exports: [PurchaseOrderService, PurchaseOrderCartService, ReorderSuggestionReader, PurchaseOrderReader],
 })
 export class ProcurementModule {}
