+++ apps/core/src/modules/inventory/procurement/services/purchase-order.service.spec.ts
@@ -1,6 +1,7 @@
 import { HttpException } from '@nestjs/common';
 import { NotFoundError } from '@app/shared';
 import { PurchaseOrderService } from './purchase-order.service';
+import { PurchaseOrderReader } from './purchase-order.reader';
 import { PurchaseOrderType } from '../dto/purchase-order.dto';
 
 /** suppliers 조회 한 건만 태우는 최소 트랜잭션 흉내. */
@@ -13,7 +14,7 @@ function serviceWithSupplier(row: { defaultWarehouseId: string | null } | undefi
     })),
   };
   const dbService = { run: jest.fn((fn: (executor: typeof trx) => unknown) => fn(trx)) };
-  return new PurchaseOrderService(dbService as never, {} as never);
+  return new PurchaseOrderService(dbService as never, {} as never, new PurchaseOrderReader(dbService as never));
 }
 
 describe('PurchaseOrderService 공급사 기본 창고', () => {
@@ -54,7 +55,7 @@ describe('PurchaseOrderService 예외 규약', () => {
       })),
     };
     const dbService = { run: jest.fn((fn: (executor: typeof trx) => unknown) => fn(trx)) };
-    return new PurchaseOrderService(dbService as never, {} as never);
+    return new PurchaseOrderService(dbService as never, {} as never, new PurchaseOrderReader(dbService as never));
   }
 
   it('없는 발주 조회는 @app/shared 의 NotFoundError 를 던진다 (Nest 예외가 아니다)', async () => {
